add genuser.sh, fix errors
macOS and Windows might not have htpasswd tools available, so adding this to the tpotinit image. users can run genuser.sh which simply contains a docker command to run tpotinit to create a user and add it to the T-Pot config (.env).
Fix an issue where WEB_USER was added with hyphens
Fix issues where shebang was incorrect
Update .env / env.example